requirements.txt
fastapi[standard]
pymongo
python-dotenv
cloudinary
uvicorn